requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
class-registry==2.1.2
Django==3.2
django-admin==2.0.2
django-decouple==2.1
django-excel-response2==3.0.5
django-filter==23.3
django-jazzmin==2.6.0
django-restframework==0.0.1
django-six==1.0.5
djangorestframework==3.14.0
drf-yasg==1.21.7
excel-base==1.0.4
filters==1.3.2
inflection==0.5.1
install==1.3.5
isoweek==1.3.3
packaging==23.2
PyMySQL==1.1.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
regex==2023.10.3
screen==1.0.1
six==1.16.0
sqlparse==0.4.4
TimeConvert==3.0.13
tqdm==4.66.1
typing_extensions==4.7.1
tzlocal==5.0.1
uritemplate==4.1.1
xlwt==1.3.0
